Moved test helper functions into test file
